# microservices/csitOasValidationApi/csit_validation/models/validation_response.py
from __future__ import annotations
import pprint
import json




from datetime import datetime
from pydantic import BaseModel, Field, StrictBool, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List
from csit_validation.models.result import Result
from csit_validation.models.validation_response_summary import ValidationResponseSummary
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class ValidationResponse(BaseModel):
    """
    ValidationResponse
    """
    valid: StrictBool = Field(description="True if no errors were found (warnings may still exist)")
    version: StrictStr = Field(description="The version (tag) used for validation")
    ruleset: StrictStr = Field(description="The ruleset file path used for validation")
    duration_ms: StrictInt = Field(description="Validation duration in milliseconds", alias="durationMs", title="Duration Ms")
    summary: ValidationResponseSummary
    results: List[Result] = Field(description="Detailed validation results")
    validated_at: datetime = Field(description="Timestamp when validation completed", alias="validatedAt", title="Validated At")
    __properties: ClassVar[List[str]] = ["valid", "version", "ruleset", "durationMs", "summary", "results", "validatedAt"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ValidationResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of summary
        if self.summary:
            _dict['summary'] = self.summary.to_dict()
        # override the default output from pydantic by calling `to_dict()` of each item in results (list)
        _items = []
        if self.results:
            for _item in self.results:
                if _item:
                    _items.append(_item.to_dict())
            _dict['results'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ValidationResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "valid": obj.get("valid"),
            "version": obj.get("version"),
            "ruleset": obj.get("ruleset"),
            "durationMs": obj.get("durationMs"),
            "summary": ValidationResponseSummary.from_dict(obj.get("summary")) if obj.get("summary") is not None else None,
            "results": [Result.from_dict(_item) for _item in obj.get("results")] if obj.get("results") is not None else None,
            "validatedAt": obj.get("validatedAt")
        })
        return _obj